// src/models/tutor.ts
export interface Tutor{
    id?: string;
    emailId?: string;
    firstName?: string;
    lastName?: string;
    gender?: string;
    mobileNumber?: string;
    whatsappNumber?: number;
    collegeName?: string;
    department?: string;
    currentSemester?: string;
    yearOfPassing?: number;
    cgpa?: number;
    softwareSubjects?: string;
    mathsSubjects?: string;
    otherSubjects?: string;
    departmentSubjects?: string;
    departmentOfIntrest?: string;
    whyDoYouWantToJoin?: string;
    query?: string;
    solution?: string;
    bankAccountNumber?: string;
    iFSCCode?: string;
    pANNumber?: string;
    dateOfBirth?: number;
    subjectAssigned?: string;
    availability?: string;
    applicationStatus?: string;
    rejectedReason?: string;
    feedbackForAssignment?: string;
    booksShared?: boolean;
    bankAdded?: string;
    tutorName?: string;
    dateOfJoining?: number;
    dateTillWorked?: number;
    cvSubmitted?: boolean;
    dateApplied?: number;
    numberOfCallsDone?: string;
    experience?: string;
    profilePic?: string;
    degree?: string;
    initialRating?: number;
    groupCreated?: boolean;
    interviewedOn?: number;
}